requirements.txt

# test deps
pyyaml
pytest
pytest-cov
ipdb

# vscode
autopep8
